+++ src/os2datascanner/projects/report/reportapp/views/base_views.py
@@ -1,0 +1,45 @@
+# Part of the OSdatascanner system, copyright © 2014-2026 Magenta ApS.
+# This Source Code Form is subject to the terms of the Mozilla Public License,
+# v. 2.0. If a copy of the MPL was not distributed with this file, you can
+# obtain one at http://mozilla.org/MPL/2.0/.
+
+from django.contrib.auth.mixins import LoginRequiredMixin
+from django.http import HttpResponse, HttpResponseBadRequest
+from django.views.generic import View, ListView
+
+from ..models.documentreport import DocumentReport
+
+
+class HTMXEndpointView(LoginRequiredMixin, View):
+    """A view for sending POST-requests via HTMX to the backend."""
+
+    model = DocumentReport
+
+    def post(self, request, *args, **kwargs):
+
+        # Add a header value to the response before returning to initiate reload of some elements.
+        response = HttpResponse()
+        response.headers["HX-Trigger"] = "reload-htmx"
+
+        return response
+
+    def dispatch(self, request, *args, **kwargs):
+        self.is_htmx = request.headers.get('HX-Request')
+        if self.is_htmx == "true":
+            return super().dispatch(request, *args, **kwargs)
+        else:
+            return HttpResponseBadRequest("HTMX endpoint called from non-HTMX source!")
+
+
+class BaseMassView(ListView):
+    """
+    Base class for Mass-related views.
+    A ListView, defining get_queryset, returning queryset of items checked in table-checkbox.
+    Not for any use alone.
+    """
+
+    def get_queryset(self):
+        qs = super().get_queryset()
+        pks = self.request.POST.getlist("table-checkbox", [])
+        reports = qs.filter(pk__in=pks)
+        return reports